dags/video_streaming/benchmark/requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.9
pymongo==4.6.0
plotly==5.18.0
pandas==2.1.3
python-dotenv==1.0.0
faker==20.1.0